move ftl.gc.android files to client layer
