+++ test_runner/src/main/kotlin/ftl/adapter/GoogleTestMatrixAndroid.kt
@@ -3,7 +3,7 @@ package ftl.adapter
 import ftl.adapter.google.toApiModel
 import ftl.api.TestMatrix
 import ftl.api.TestMatrixAndroid
-import ftl.client.google.executeAndroidTests
+import ftl.client.google.run.android.executeAndroidTests
 import kotlinx.coroutines.runBlocking
 import com.google.testing.model.TestMatrix as GoogleTestMatrix
