--- test_runner/src/main/kotlin/ftl/client/google/ClientInfoDetailList.kt
@@ -1,9 +1,0 @@
-package ftl.client.google
-
-import com.google.testing.model.ClientInfoDetail
-
-internal fun Map<String, String>.toClientInfoDetailList() = map { (key, value) ->
-    ClientInfoDetail()
-        .setKey(key)
-        .setValue(value)
-}
