+++ test_runner/src/main/kotlin/ftl/client/google/run/android/CreateAndroidLoopTest.kt
@@ -1,4 +1,4 @@
-package ftl.gc.android
+package ftl.client.google.run.android
 
 import com.google.testing.model.AndroidTestLoop
 import com.google.testing.model.FileReference
