+++ test_runner/src/main/kotlin/ftl/client/google/run/android/GcAndroidDevice.kt
@@ -1,4 +1,4 @@
-package ftl.gc
+package ftl.client.google.run.android
 
 import com.google.testing.model.AndroidDevice
 import com.google.testing.model.AndroidDeviceList
