+++ test_runner/src/test/kotlin/ftl/api/TestMatrixTest.kt
@@ -11,9 +11,9 @@ import com.google.testing.model.TestSpecification
 import com.google.testing.model.ToolResultsExecution
 import com.google.testing.model.ToolResultsStep
 import ftl.adapter.google.toApiModel
+import ftl.client.google.run.android.GcAndroidDevice
 import ftl.config.Device
 import ftl.domain.testmatrix.updateWithMatrix
-import ftl.gc.GcAndroidDevice
 import ftl.json.createAndUpdateMatrix
 import ftl.reports.outcome.make
 import ftl.test.util.FlankTestRunner
